+++ halotools/empirical_models/phase_space_models/__init__.py
@@ -1,5 +1,6 @@
 from .nfw_phase_space import *
 from .trivial_phase_space import *
+from .monte_carlo_helpers import *
 
 from .profile_models import *
 from .velocity_models import *
