+++ halotools/empirical_models/phase_space_models/monte_carlo_helpers.py
